+++ application/models/Fleet.php
@@ -2,6 +2,7 @@
 class Fleet extends CSV_Model 
 {
     
+
     /*
     var $data = array
         (
@@ -71,22 +72,75 @@ class Fleet extends CSV_Model
         parent::__construct(APPPATH . '../data/fleet.csv', 'id');
     }
 
+   /* public function add($fleet)
+    {
+        
+       array_push($this->data, $fleet);
+        // inject each "record" key into the record itself, for ease of presentation
+        foreach ($this->data as $key => $record)
+        {
+            $record['key'] = $key;
+            $this->data[$key] = $record;
+        }
+    }
+    */
+    /*
+    public function update($fleet)
+    {
+        $found = 0;
+        foreach($this->data  as $datafleet) {
+            if ($datafleet['id'] == $fleet['id']) {
+                $datafleet = $fleet;
+                $found = 1;
+            }
+        }
+        if (found == 0) {
+            add($fleet);
+        }
+    }
+    */
+    // provide form validation rules
+    public function rules()
+    {
+        $config = array(
+            ['field' => 'id', 'label' => 'Fleet Id', 'rules' => 'alpha_numeric_spaces|max_length[64]'],
+            ['field' => 'plane_id', 'label' => 'Plane Id', 'rules' => 'alpha_numeric_spaces|max_length[64]'],
+            ['field' => 'model', 'label' => 'Model', 'rules' => 'alpha_numeric_spaces|max_length[64]'],
+            ['field' => 'price', 'label' => 'Price', 'rules' => 'alpha_numeric_spaces|max_length[64]'],
+            ['field' => 'seats', 'label' => 'Seats', 'rules' => 'alpha_numeric_spaces|max_length[64]'],
+            ['field' => 'reach', 'label' => 'Reach', 'rules' => 'alpha_numeric_spaces|max_length[64]'],
+            ['field' => 'cruise', 'label' => 'Cruise', 'rules' => 'alpha_numeric_spaces|max_length[64]'],
+            ['field' => 'takeoff', 'label' => 'Takeoff', 'rules' => 'alpha_numeric_spaces|max_length[64]'],
+            ['field' => 'hourly', 'label' => 'Hourly', 'rules' => 'alpha_numeric_spaces|max_length[64]'],
+        );
+        return $config;
+    }
+    
+    
+    
+    
+    
+
     /*
     // retrieve a single quote, null if not found
     public function get($which)
     {
         return !isset($this->data[$which]) ? null : $this->data[$which];
     }
-    public function getPlane($id) {
-       foreach($this->data  as $fleet) {
+     */
+    
+    
+    public function getFleet($id) {
+       foreach($this->allAsArray()  as $fleet) {
             if ($fleet['id'] == $id) {
                 return $fleet;
             }
-            
         }
         return null;
     }
-     */
+    
+    
+    
     // retrieve all of the quotes
     
     public function allAsArray()
@@ -96,6 +150,8 @@ class Fleet extends CSV_Model
            $fleetArray[$plane->id] = (array)$plane;
         return $fleetArray;
     }
+    
+
 
     /**
      * This function is not needed, use size() instead