+++ application/views/fleet.php
@@ -1,5 +1,7 @@
 <h1>{title}</h1>
-{thetable}
+<form role="form" action="/fleet/submit" method="post">
+    {thetable}
+</form>
 {jsonbutton}